==> experiments/02_geo_data/02_poalrs_st.ipynb <==
import pyogrio

import polars as pl
import polars_st as st

import pytz

df = pl.Series(
    "geometry",
    [
        "POINT(0 0)",
        "POINT(1 2)",
    ],
).to_frame()
df.select(st.from_wkt("geometry")).st.to_geopandas().explore()

import polars as pl
import polars_st as st

df = pl.DataFrame(
    {
        "geometry": [
            "POLYGON ((0 0, 0 1, 1 1, 1 0, 0 0))",
            "POLYGON ((0 0, 0 1, 1 1, 0 0))",
        ]
    }
)
gdf = df.select(geometry=st.from_wkt("geometry"))
area = gdf.select(pl.col("geometry").st.area())
area

gdf.select(pl.col("geometry").st.centroid()).st.to_geopandas()

gdf.st.to_geopandas().explore()
