==> experiments/02_geo_data/02.6_intersection_changes.ipynb <==
%load_ext autoreload
%autoreload 2

import geopandas as gpd
import pandas as pd
import polars as pl
import polars_st as st
import pyogrio
from electoralyze import region
from electoralyze.common import constants
from electoralyze.common.geometry import to_geopandas, to_geopolars

sa1 = region.SA1_2021.get_raw_geometry()

list(region.SA1_2021.get_ids())[10000:10003]

cropped_list = sa1.filter(pl.col(region.SA1_2021.id).cast(pl.String).str.starts_with(12202)).with_columns(
    pl.col(region.SA1_2021.id).cast(pl.String).str.slice(0, 9).alias("sa2")
)
cropped_list  # .pipe(to_geopandas).explore()

cropped_list.group_by("sa2").agg(pl.col("geometry").st.union_all()).sort("sa2")

cropped_list.with_columns(st.geom("geometry").st.union_all())

st.GeoDataFrame(
    {
        "x": [1],
        "geometry": sa1.filter(pl.col(region.SA1_2021.id).cast(pl.String).str.starts_with(122021691))[
            "geometry"
        ].st.union_all(),
    }
)  # .pipe(to_geopandas)#.explore()
