==> experiments/02_geo_data/03_overlap.ipynb <==
import polars as pl
import polars_st as st

bottom_tiles = pl.DataFrame(
    {
        "top": [
            "POLYGON ((0 0, 0 1, -1 1, -1 0, 0 0))",
            "POLYGON ((0 0, 0 1, 1 1, 1 0, 0 0))",
        ]
    }
).select(top=st.from_wkt("top"))

bottom_tiles.rename({"top": "geometry"}).st.to_geopandas().explore()

top_tiles = pl.DataFrame(
    {
        "bottom": [
            "POLYGON ((0.5 0, 0.5 1, -0.5 1, -0.5 0, 0.5 0))",
        ]
    }
).select(bottom=st.from_wkt("bottom"))

top_tiles.rename({"bottom": "geometry"}).st.to_geopandas().explore()

pl.concat(
    [top_tiles.rename({"bottom": "geometry"}), bottom_tiles.rename({"top": "geometry"})]
).st.to_geopandas().explore()

intersection = bottom_tiles.join(top_tiles, how="cross").select(
    pl.col("top").st.intersection(pl.col("bottom")).alias("geometry")
)
intersection.st.to_geopandas().explore()

intersection.select(st.geom("geometry").st.area())

# more complex geometry

bottom_tiles = pl.DataFrame(
    {
        "top_name": ["M", "N", "O", "P"],
        "top": [
            "POLYGON ((0 0, 0 1, -1 1, -1 0, 0 0))",
            "POLYGON ((0 0, 0 1, 1 1, 1 0, 0 0))",
            "POLYGON ((0 0, 0 -1, -1 -1, -1 0, 0 0))",
            "POLYGON ((0 0, 0 -1, 1 -1, 1 0, 0 0))",
        ],
    }
).with_columns(top=st.from_wkt("top"))

top_tiles = pl.DataFrame(
    {
        "bottom_name": ["A", "B", "C"],
        "bottom": [
            # "POLYGON ((-1 0, 0 1, 1 0, 0 -1, -1 0))",
            "POLYGON ((-1 -1, 0 1, 1 -1, -1 -1))",
            "POLYGON ((-1 -1, 0 1, -2 1, -1 -1))",
            "POLYGON ((1 -1, 0 1, 2 1, 1 -1))",
        ],
    }
).with_columns(bottom=st.from_wkt("bottom"))

intersection = bottom_tiles.join(top_tiles, how="cross").select(
    pl.exclude("top", "bottom"), pl.col("top").st.intersection(pl.col("bottom")).alias("geometry")
)
pl.concat(
    [top_tiles[["bottom"]].rename({"bottom": "geometry"}), bottom_tiles[["top"]].rename({"top": "geometry"})]
).st.to_geopandas().explore()

intersection.filter(st.geom("geometry").st.area() != 0).with_columns(st.geom("geometry").st.to_wkt()).to_dict(
    as_series=False
)

intersection.filter(st.geom("geometry").st.area() != 0).st.to_geopandas().explore()

intersection_area = (
    intersection.filter(st.geom("geometry").st.area() != 0)
    .with_columns(pl.col("geometry").st.area().alias("area"))
    .drop("geometry")
)
intersection_area

# %%timeit
(
    intersection_area.with_columns(
        pl.col("area").sum().over("top_name").alias("top_area"),
        pl.col("area").sum().over("bottom_name").alias("bottom_area"),
    )
)
