==> experiments/02_geo_data/3.2_redistribute.ipynb <==
%load_ext autoreload
%autoreload 2

import geopandas as gpd
import pandas as pd
import polars as pl
import polars_st as st
import pyogrio

# from electoralyze import region``
from electoralyze.common import constants
from electoralyze.common.geometry import to_geopandas, to_geopolars
from electoralyze.common.testing.region_fixture import (
    RegionMocked,
    create_fake_regions,
    read_true_geometry,
    read_true_metadata,
)
from electoralyze.region.redistribute import redistribute
from electoralyze.region.redistribute.mapping import (
    _create_intersection_area_mapping,
    _get_remaining_area,
    create_region_mapping_base,
)
from polars import testing

intersection_area = _create_intersection_area_mapping(
    geometry_from=read_true_geometry("triangle"),
    geometry_to=read_true_geometry("square"),
)
intersection_area.pipe(print)

import shutil
import tempfile

temp_dir = tempfile.TemporaryDirectory(delete=False)
region_ = create_fake_regions(temp_dir.name)
region_
region_.quadrant.geometry
region_.triangle.geometry
region_.rectangle.geometry
region_.square.geometry
region_.l_and_r.geometry
shutil.rmtree(temp_dir.name)

from electoralyze.common.testing.region_fixture import ONE_SQUARE_REGION_ID

RegionMocked

data_by_from = pl.DataFrame({"triangle": ["A"], "data": [1]})
print(data_by_from)
redistribute(
    data_by_from,
    region_from=region_.triangle,
    region_to=region_.quadrant,
    mapping="intersection_area",
    redistribute_with_full=False,
)

data_by_from = pl.DataFrame({"rectangle": ["X", "Y", "Z"], "data": [1, 1, 1]})
print(data_by_from)
redistribute(
    data_by_from,
    region_from=region_.rectangle,
    region_via=region_.quadrant,
    region_to=region_.triangle,
    mapping="intersection_area",
    redistribute_with_full=False,
)

data_by_from = pl.DataFrame({"square": ["main"], "data": [100]})
print(data_by_from)
redistribute(
    data_by_from,
    region_from=region_.square,
    region_via=region_.quadrant,
    region_to=region_.l_and_r,
    mapping="intersection_area",
    redistribute_with_full=False,
)

data_by_from = pl.DataFrame({"square": ["main"], "data": [100]})
print(data_by_from)
redistribute(
    data_by_from,
    region_from=region_.square,
    region_via=region_.triangle,
    region_to=region_.l_and_r,
    mapping="intersection_area",
    redistribute_with_full=False,
)

def test_redistribute_via(
    region: RegionMocked, _name: str, region_ids: dict, redistribute_kwargs: dict, redistributed_expected: dict
):
    redistributed_to = redistribute(
        region_from=region.from_id(region_ids["region_id_from"]),
        region_to=region.from_id(region_ids["region_id_via"]),
        **redistribute_kwargs,
    )

    pl.testing.assert_frame_equal(
        redistributed_to,
        redistributed_expected["to"],
        check_row_order=False,
        check_column_order=False,
    )

    redistributed_via = redistribute(
        region_from=region.from_id(region_ids["region_id_from"]),
        region_via=region.from_id(region_ids["region_id_via"]),
        region_to=region.from_id(region_ids["region_id_to"]),
        **redistribute_kwargs,
    )

    pl.testing.assert_frame_equal(
        redistributed_via,
        redistributed_expected["via"],
        check_row_order=False,
        check_column_order=False,
    )

pl.DataFrame({"l_and_r": ["L", "R"], "data": [32, 64]}).rows(named=True)

data_by_from = pl.DataFrame({"l_and_r": ["L", "R"], "data": [32, 64]})
print(data_by_from)
_name = "L R through Triangle to quadrents"
redistribute_kwargs = dict(
    data_by_from=data_by_from,
    mapping="intersection_area",
    redistribute_with_full=False,
)
region_ids = dict(
    region_id_from="l_and_r",
    region_id_to="quadrant",
    region_id_via="triangle",
)

redistributed_expected = dict(
    via=pl.DataFrame({"quadrant": ["O", "M", "P", None, "N"], "data": [20.0, 12.0, 22.0, 24.0, 18.0]}),
    to=pl.DataFrame({"triangle": ["A", "B", "C"], "data": [48.0, 16.0, 32.0]}),
)

test_redistribute_via(
    _name=_name,
    region=region_,
    region_ids=region_ids,
    redistribute_kwargs=redistribute_kwargs,
    redistributed_expected=redistributed_expected,
)

redistribute(
    region_from=region_.from_id(region_ids["region_id_from"]),
    region_via=region_.from_id(region_ids["region_id_via"]),
    region_to=region_.from_id(region_ids["region_id_to"]),
    **redistribute_kwargs,
)  # .to_dict(as_series=False)
# ).rows(named = True)

pl.DataFrame(
    {
        ONE_SQUARE_REGION_ID: ["main"],
        "data": [64.0],
    }
).pipe(print)

redistribute(
    data_by_from=pl.DataFrame(
        {
            ONE_SQUARE_REGION_ID: ["main"],
            "data": [64.0],
        }
    ),
    region_from=region_.square,
    region_via=region_.triangle,
    region_to=region_.quadrant,
    mapping="intersection_area",
    redistribute_with_full=False,
    # aggregation="min",
    errors=False,
).pipe(print)  # .to_dict(as_series=False)
# ).rows(named = True)

# region_.quadrant.geometry.pipe(to_geopandas).pipe(print)
# region_.square.geometry.pipe(to_geopandas).pipe(print)
# region_.triangle.geometry.pipe(to_geopandas).pipe(print)
# region_.quadrant.geometry.pipe(print)
# region_.square.geometry.pipe(print)
# region_.triangle.geometry.pipe(print)

redistribute_kwargs["data_by_from"].pipe(print)

redistributed_expected.get("to").rows(named=True)

from electoralyze.common.testing.region_fixture import FOUR_SQUARE_REGION_ID, LEFT_RIGHT_REGION_ID, ONE_SQUARE_REGION_ID

_name = "from one square -> via quadrants -> to L and R, "
region_ids = dict(
    region_id_from=ONE_SQUARE_REGION_ID,
    region_id_via=FOUR_SQUARE_REGION_ID,
    region_id_to=LEFT_RIGHT_REGION_ID,
)
redistribute_kwargs = dict(
    data_by_from=pl.DataFrame(
        [
            {ONE_SQUARE_REGION_ID: "O", "data": 100},
        ]
    ),
    mapping="intersection_area",
    redistribute_with_full=True,
)
redistributed_expected = dict(
    to=pl.DataFrame(
        [
            {FOUR_SQUARE_REGION_ID: "N", "data": 25.0},
            {FOUR_SQUARE_REGION_ID: "M", "data": 25.0},
            {FOUR_SQUARE_REGION_ID: "O", "data": 25.0},
            {FOUR_SQUARE_REGION_ID: "P", "data": 25.0},
        ]
    ),
    via=pl.DataFrame(
        [
            {LEFT_RIGHT_REGION_ID: "L", "data": 50.0},
            {LEFT_RIGHT_REGION_ID: "R", "data": 50.0},
        ]
    ),
)

_create_intersection_area_mapping(
    region_.from_id(region_ids["region_id_from"]).geometry,
    region_.from_id(region_ids["region_id_via"]).geometry,
).pipe(print)

# region_.from_id(region_ids["region_id_from"]).geometry.pipe(to_geopandas).explore()

# region_.from_id(region_ids["region_id_via"]).geometry.pipe(to_geopandas).explore()

redistribute(
    region_from=region_.from_id(region_ids["region_id_from"]),
    # region_via=region_.from_id(region_via_id),
    region_to=region_.from_id(region_ids["region_id_via"]),
    # region_to=region_.from_id(region_to_id),
    data_by_from=redistribute_kwargs["data_by_from"],
    mapping=redistribute_kwargs["mapping"],
    redistribute_with_full=False,
    errors="warning",
)

from polars.exceptions import ColumnNotFoundError

try:
    region_.from_id(ONE_SQUARE_REGION_ID).geometry.select("square1")
except ColumnNotFoundError:
    print("das")

redistribute(
    region_from=region_.from_id(ONE_SQUARE_REGION_ID),
    region_to=region_.from_id(ONE_SQUARE_REGION_ID),
    data_by_from=pl.DataFrame({ONE_SQUARE_REGION_ID: ["O"], "data": [100]}),
    # create_region_mapping_base=False,
    redistribute_with_full=False,
)

test_redistribute_via(
    _name=_name,
    region=region_,
    region_ids=region_ids,
    redistribute_kwargs=redistribute_kwargs,
    redistributed_expected=redistributed_expected,
)
