==> experiments/02_geo_data/03.5_same_geom_overlap.ipynb <==
import geopandas as gpd

from shapely.geometry import box

# Create a square geometry (from coordinates 0,0 to 1,1)
square = box(0, 0, 1, 1)

# Create GeoDataFrame with the square
gdf = gpd.GeoDataFrame(geometry=[square])
gdf_ = gdf.merge(gdf, how="cross")
print(gdf_)
gdf_.assign(area=lambda df: df["geometry_x"].intersects(df["geometry_y"]))
