==> experiments/02_geo_data/05_read_from_cloud.ipynb <==
import os

import geopandas as gpd
import polars as pl
import polars_st as st
import pyogrio
import requests
from electoralyze.common.constants import ROOT_DIR

raw_data_url = "https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/digital-boundary-files/SA1_2021_AUST_SHP_GDA2020.zip"
raw_data_file = f"{ROOT_DIR}/data/raw/temp/sa1.zip"

from electoralyze.common.files import create_path, download_file

create_path(raw_data_file)

download_file(raw_data_url, raw_data_file)

pyogrio.read_dataframe(raw_data_file)

raw_data_url = "https://raw.githubusercontent.com/datasets/geo-boundaries-world-110m/master/countries.geojson"
raw_data_file = f"{ROOT_DIR}/data/raw/temp/countries.geojson"
download_file(raw_data_url, raw_data_file)

from electoralyze.common.geometry import to_geopolars

geometry_raw = pyogrio.read_dataframe(raw_data_file).pipe(to_geopolars)
geometry_raw
