==> experiments/02_geo_data/06_federal_election_2022_boundaries.ipynb <==
import os

import geopandas as gpd
import polars as pl
import polars_st as st
import pyogrio
import requests
from electoralyze import region
from electoralyze.common.constants import ROOT_DIR
from electoralyze.common.geometry import to_geopandas, to_geopolars

region.Federal2022.download_data()

geometry_raw = region.Federal2022._get_geometry_raw()

geometry_raw

cls = region.Federal2022
geometry_raw.select(
    pl.col("Elect_div").alias(cls.id),
    pl.struct(
        pl.col("Elect_div").alias(cls.name),
        pl.col("Sortname").alias("short_name"),
        pl.col("Area_SqKm").alias("area_sqkm"),
        pl.col("Actual").alias("actual"),
    ).alias("metadata"),
    pl.col("geometry").st.to_wkt().alias("geometry"),
).pipe(print)

region.Federal2022.process_raw(force_new=True)

region.Federal2022.geometry.pipe(to_geopandas).explore()
